Bump to 1.20.5

+++ src/main/java/nota/utils/TimeUtils.java
@@ -49,21 +49,21 @@ public class TimeUtils {
 		long milisTotal = (long) ((ticks / speed) * 1000);
 
 		int hours = 0;
-		if(time.contains("h")) {
+		if (time.contains("h")) {
 			hours = (int) Math.floor(milisTotal / 1000 / 60 / 60);
-			milisTotal -= hours * 1000 * 60 * 60;
+			milisTotal -= (long) hours * 1000 * 60 * 60;
 		}
 
 		int minutes = 0;
-		if(time.contains("m")) {
+		if (time.contains("m")) {
 			minutes = (int) Math.floor(milisTotal / 1000 / 60);
-			milisTotal -= minutes * 1000 * 60;
+			milisTotal -= (long) minutes * 1000 * 60;
 		}
 
 		int seconds = 0;
-		if(time.contains("s")) {
+		if (time.contains("s")) {
 			seconds = (int) Math.floor(milisTotal / 1000);
-			milisTotal -= seconds * 1000;
+			milisTotal -= seconds * 1000L;
 		}
 
 		time = time.replace("hh", String.format("%02", hours));
